std/testing: renamed from test
