--- examples/gc.garden
@@ -1,23 +1,0 @@
-# vim: filetype=python
-import ("io" "runtime/gc" "string")
-
-let combination = string.concat("Hello" " " "World")
-io.println(combination)
-# manual gc run to check if reachability is correct
-gc.cycle()
-# second gc run, just to stress everything
-gc.cycle()
-
-let combination = string.concat(combination "|" combination "|" combination)
-io.println(combination)
-gc.cycle()
-
-# new allocation
-let combination = string.append(combination combination)
-fn child_scope {
-    # in this scope the gc needs to be able to mark the parent scope
-    gc.cycle()
-}
-child_scope()
-
-io.println(combination)
